Share plugin extract path; gate pin bumps on version change
Add plugin_catalog.extract_plugin (manifest version + components) used by
index generation and the SHA bump bot. Index now records version; bumps
only open when HEAD moved and version at HEAD differs from version at the
current pin (pin HEAD at discovery time).
